refactor: #78 size 토큰 파일 width로 이름 변경
