+++ ds/components/molecules/workout/Workout.tsx
@@ -3,7 +3,7 @@ import { WorkoutProps, WorkoutSetData, WorkoutSetDataField, WorkoutType } from "
 import { workoutVariants } from "@/ds/styles/tokens/workout/variants"
 import { gridColumns, GridColumn } from "@/ds/styles/tokens/workout/gridColumns"
 import { H2 } from "@/ds/components/atoms/text/TextWrapper"
-import { workoutWidths } from "@/ds/styles/tokens/workout/sizes"
+import { workoutWidths } from "@/ds/styles/tokens/workout/width"
 import { workoutFields } from "@/ds/styles/tokens/workout/fields"
 import { Input } from "@/ds/components/atoms/input/Input"
 import Text from "@/ds/components/atoms/text/Text"
